requirements.txt
asgiref==3.7.1
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2.1
django-cleanup==7.0.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-utilitas==1.3.9
djangorestframework==3.14.0
drf-flex-fields==1.0.2
drf-yasg==1.21.5
filelock==3.12.0
gunicorn==20.1.0
identify==2.5.24
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pre-commit==3.3.2
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
